// src/app/interfaces/product-model.ts
/* eslint-disable @typescript-eslint/naming-convention */
// Generated by https://quicktype.io

export interface ProductModel {
  id: number;
  nombre: string;
  clasificador_id: string;
  codigo_agrupador: string;
  codigo: string;
  codigo_secundario: string;
  codigo_barra: string;
  descripcion: string;
  departamentos: string;
  estatus: boolean;
  fecha_hora: string;
  relacion_sat: string;
  requiere_serie: boolean;
  tipo: string;
  venta_fraccionaria: boolean;
  ancho: number;
  alto: number;
  largo: string;
  peso: string;
  unidad_principal_id: string;
  numero_ventas: number;
  estatus_id: number;

  url_amigable: string;
  bloqueado: boolean;
  marca: string;
  objeto_imp: string;
}